+++ src/main.c
@@ -86,8 +86,8 @@ int main(int argc, char *argv[]) {
     LLL(B, N);
 
     GS_Info *gs_info = gram_schmidt(B, N);
-    long double bound = lambda_1(gs_info->Bs, N);
-    // printf("Bound: %.6Lf\n", bound);
+    double bound = lambda_1(gs_info->Bs, N);
+    // printf("Bound: %.6f\n", bound);
 
     double result = schorr_euchner(N, gs_info, bound);
     // printf("%8.8f\n", result);
